backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Supabase
supabase>=2.7.0

# OCR (easyocr removed: optional, requires torch 2GB+, incompatible with Python 3.13)
pytesseract==0.3.10
pdf2image==1.16.3
Pillow>=10.4.0

# Image Preprocessing
opencv-python>=4.10.0
numpy>=2.0.0

# ML Utilities (spacy removed: Sprint 4 feature, not yet implemented)
scikit-learn>=1.5.0
pandas>=2.2.2

# Utilities
python-dotenv==1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# For PDF processing
PyMuPDF>=1.24.0
